requirements.txt
pandas>=2.2.3
plotly>=5.24.1
psycopg2-binary>=2.9.10
sqlalchemy>=2.0.37
streamlit>=1.41.1
bcrypt>=4.1.2
python-dotenv>=1.0.1
alembic>=1.13.1
matplotlib>=3.8.3
numpy>=1.26.4
pillow>=10.2.0
watchdog>=4.0.0
pytest>=8.0.2
pytest-cov>=4.1.0
black>=24.2.0
flake8>=7.0.0
isort>=5.13.2
groq>=0.4.1